requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.6
distro==1.9.0
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
filelock==3.25.2
fsspec==2026.3.0
google-auth==2.49.1
google-genai==1.69.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.6
numba==0.64.0
numpy==2.4.4
openai-whisper==20250625
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
python-dotenv==1.2.2
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
regex==2026.3.32
requests==2.33.1
setuptools==81.0.0
sniffio==1.3.1
sqlparse==0.5.5
sympy==1.14.0
tenacity==9.1.4
tiktoken==0.12.0
torch==2.11.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
yt-dlp==2026.3.17
